stdio working with right tools
