+++ mcp_template/core/tool_manager.py
@@ -301,10 +301,16 @@ class ToolManager:
             mcp_host = os.getenv("MCP_HOST", "127.0.0.1")
             endpoint = f"http://{mcp_host}:{external_port}/mcp"
 
-            # Run async MCP connection
-            return asyncio.run(
-                self._async_discover_via_http(endpoint, timeout, deployment)
-            )
+            # Use BaseProbe's HTTP discovery methods
+            # Create a temporary probe instance to access the HTTP discovery
+            if self.backend_type == "kubernetes":
+                probe = KubernetesProbe()
+            elif self.backend_type == "docker":
+                probe = DockerProbe()
+            else:
+                raise ValueError("Only docker and kuberenetes backends are supported")
+
+            return asyncio.run(probe._async_discover_via_http(endpoint, timeout))
 
         except Exception as e:
             logger.debug(f"Failed to discover from running deployment: {e}")
@@ -465,12 +471,9 @@ class ToolManager:
         self, template_or_deployment: str, timeout: int
     ) -> List[Dict]:
         """Helper method to discover tools via MCP JSON-RPC (for already deployed services)."""
-
         try:
             # Get deployment manager to find running deployments
-            deployment_manager = self.deployment_managers.get(self.backend_type)
-            if not deployment_manager:
-                return []
+            deployment_manager = DeploymentManager(self.backend_type)
 
             # Find deployments for this template
             deployments = deployment_manager.find_deployments_by_criteria(
@@ -498,211 +501,18 @@ class ToolManager:
             mcp_host = os.getenv("MCP_HOST", "127.0.0.1")
             endpoint = f"http://{mcp_host}:{external_port}/mcp"
 
-            # Run async MCP connection
-            return asyncio.run(
-                self._async_discover_via_http(endpoint, timeout, deployment)
-            )
+            # Use BaseProbe's HTTP discovery methods
+            if self.backend_type == "kubernetes":
+                probe = KubernetesProbe()
+            else:
+                probe = DockerProbe()
+
+            return asyncio.run(probe._async_discover_via_http(endpoint, timeout))
 
         except Exception as e:
             logger.debug(
                 f"MCP JSON-RPC discovery failed for {template_or_deployment}: {e}"
             )
-            return []
-
-    async def _async_discover_via_http(
-        self, endpoint: str, timeout: int, deployment: Dict
-    ) -> List[Dict]:
-        """Async MCP JSON-RPC discovery with proper session management."""
-        try:
-
-            # Try multiple MCP connection approaches
-            tools = []
-
-            # Approach 1: Direct tools/list call
-            tools = await self._try_direct_tools_list(endpoint, timeout)
-            if tools:
-                logger.info(
-                    f"Discovered {len(tools)} tools via direct tools/list from deployment {deployment.get('name', 'unknown')}"
-                )
-                return tools
-
-            # Approach 2: Full MCP handshake (like VS Code does)
-            tools = await self._try_mcp_handshake(endpoint, timeout)
-            if tools:
-                logger.info(
-                    f"Discovered {len(tools)} tools via MCP handshake from deployment {deployment.get('name', 'unknown')}"
-                )
-                return tools
-
-            # Approach 3: WebSocket connection (if HTTP fails)
-            ws_endpoint = endpoint.replace("http://", "ws://").replace(
-                "/mcp", "/mcp/ws"
-            )
-            tools = await self._try_websocket_connection(ws_endpoint, timeout)
-            if tools:
-                logger.info(
-                    f"Discovered {len(tools)} tools via WebSocket from deployment {deployment.get('name', 'unknown')}"
-                )
-                return tools
-
-            return []
-
-        except Exception as e:
-            logger.debug(f"Async MCP discovery failed: {e}")
-            return []
-
-    async def _try_direct_tools_list(self, endpoint: str, timeout: int) -> List[Dict]:
-        """Try direct tools/list MCP call."""
-        try:
-
-            async with aiohttp.ClientSession(
-                timeout=aiohttp.ClientTimeout(total=timeout)
-            ) as session:
-                payload = {
-                    "jsonrpc": "2.0",
-                    "id": 1,
-                    "method": "tools/list",
-                    "params": {},
-                }
-
-                logger.debug(f"Trying direct tools/list call to {endpoint}")
-                async with session.post(
-                    endpoint, json=payload, headers={"Content-Type": "application/json"}
-                ) as response:
-                    if response.status == 200:
-                        result = await response.json()
-                        if result.get("result") and "tools" in result["result"]:
-                            return result["result"]["tools"]
-            return []
-        except Exception as e:
-            logger.debug(f"Direct tools/list failed: {e}")
-            return []
-
-    async def _try_mcp_handshake(self, endpoint: str, timeout: int) -> List[Dict]:
-        """Try full MCP handshake like VS Code MCP clients do."""
-        try:
-            async with aiohttp.ClientSession(
-                timeout=aiohttp.ClientTimeout(total=timeout)
-            ) as session:
-                # Step 1: Initialize connection
-                init_payload = {
-                    "jsonrpc": "2.0",
-                    "id": 1,
-                    "method": "initialize",
-                    "params": {
-                        "protocolVersion": "2024-11-05",
-                        "capabilities": {
-                            "roots": {"listChanged": True},
-                            "sampling": {},
-                        },
-                        "clientInfo": {"name": "mcp-template", "version": "1.0.0"},
-                    },
-                }
-
-                logger.debug(f"Trying MCP handshake to {endpoint}")
-                async with session.post(
-                    endpoint,
-                    json=init_payload,
-                    headers={"Content-Type": "application/json"},
-                ) as response:
-                    if response.status != 200:
-                        return []
-
-                    init_result = await response.json()
-                    if init_result.get("error"):
-                        logger.debug(f"MCP initialize failed: {init_result['error']}")
-                        return []
-
-                # Step 2: Send initialized notification
-                notif_payload = {
-                    "jsonrpc": "2.0",
-                    "method": "notifications/initialized",
-                }
-
-                async with session.post(
-                    endpoint,
-                    json=notif_payload,
-                    headers={"Content-Type": "application/json"},
-                ) as response:
-                    pass  # Notification doesn't need response
-
-                # Step 3: List tools
-                tools_payload = {
-                    "jsonrpc": "2.0",
-                    "id": 2,
-                    "method": "tools/list",
-                    "params": {},
-                }
-
-                async with session.post(
-                    endpoint,
-                    json=tools_payload,
-                    headers={"Content-Type": "application/json"},
-                ) as response:
-                    if response.status == 200:
-                        result = await response.json()
-                        if result.get("result") and "tools" in result["result"]:
-                            return result["result"]["tools"]
-
-            return []
-        except Exception as e:
-            logger.debug(f"MCP handshake failed: {e}")
-            return []
-
-    async def _try_websocket_connection(
-        self, ws_endpoint: str, timeout: int
-    ) -> List[Dict]:
-        """Try WebSocket connection for MCP (some servers prefer this)."""
-        try:
-            async with aiohttp.ClientSession(
-                timeout=aiohttp.ClientTimeout(total=timeout)
-            ) as session:
-                async with session.ws_connect(ws_endpoint) as ws:
-                    # Send initialize
-                    init_msg = {
-                        "jsonrpc": "2.0",
-                        "id": 1,
-                        "method": "initialize",
-                        "params": {
-                            "protocolVersion": "2024-11-05",
-                            "capabilities": {},
-                            "clientInfo": {"name": "mcp-template", "version": "1.0.0"},
-                        },
-                    }
-
-                    await ws.send_str(json.dumps(init_msg))
-
-                    # Wait for initialize response
-                    msg = await ws.receive()
-                    if msg.type != aiohttp.WSMsgType.TEXT:
-                        return []
-
-                    # Send initialized notification
-                    notif_msg = {
-                        "jsonrpc": "2.0",
-                        "method": "notifications/initialized",
-                    }
-                    await ws.send_str(json.dumps(notif_msg))
-
-                    # Request tools
-                    tools_msg = {
-                        "jsonrpc": "2.0",
-                        "id": 2,
-                        "method": "tools/list",
-                        "params": {},
-                    }
-                    await ws.send_str(json.dumps(tools_msg))
-
-                    # Get tools response
-                    msg = await ws.receive()
-                    if msg.type == aiohttp.WSMsgType.TEXT:
-                        result = json.loads(msg.data)
-                        if result.get("result") and "tools" in result["result"]:
-                            return result["result"]["tools"]
-
-            return []
-        except Exception as e:
-            logger.debug(f"WebSocket connection failed: {e}")
             return []
 
     def _cache_tools(
